Move out functions from src/timezone/gptime.c.
Let's refrain from having any GPDB-specific functions or modifications
vs upstream in src/timezone. That way, we can easily replace the whole
directory from upstream, when we need to update the timezone data.
The code in gptime.c wasn't really related to time zones anyway.

The functions in gptime.c, and related inline functions in pgtime.h, were
only used in ic_tcp.c, so instead of keeping gptime.c as a separate file,
move the functions to ic_tcp.c itself. I don't think we want to encourage
using them more, by keeping them as a separate file.

pgtime.h is now identical to upstream. src/timezone/ isn't identical to
any particular upstream version that I can see, but it's close. I don't
see any meaningful differences there, I think it's just a mixture of
different upstream versions and different versions of the timezone code
and data. Going forward, we could and should just replace it with the
latest upstream version.
